Table Organism -> Renames table_data.py to table_cell.py
To match the class name.
