+++ automancy/elementals/organisms/table/table_cell.py
